+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Possibility to specify directory using linux path syntax
 - Possibility to specify base directory
 - Progress hook
+- Possibility to follow redirects
 
 ### Changed
 - Small refactoring